Functions have been moved to a separate file
