+++ lib/bot/handlers/hdr_generate-image.go
@@ -5,9 +5,7 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"net/url"
 	"os"
-	"path"
 	"path/filepath"
 
 	"github.com/JackBekket/hellper/lib/localai"
@@ -142,21 +140,4 @@ func getImage(imageURL, localAITokenHeader string) (string, error) {
 
 	return fileName, nil
 
-}
-
-func transformURL(inputURL string) string {
-	parsedURL, _ := url.Parse(inputURL)
-	fileName := path.Base(parsedURL.Path)
-	return fileName
-}
-
-func getURL(endpoint string, suffix string) string {
-	if suffix == "" {
-		suffix = ai_ImageGenerationSuffix
-	}
-	joined, err := url.JoinPath(endpoint, suffix)
-	if err != nil {
-		return endpoint + suffix
-	}
-	return joined
 }
